Parse received headers
The received-date doesn't work really. It may only be available when
connected to a server. Once a mail is read from a string, the javamail
method always returns null. Even if there is a Delivery-Date or
something similiar present in the string. So it may be better to
inspect the Received headers that are often there.