Introduce BodyContent class
The `String` in `MailBody` is not enough, since the original encoding
of a message is lost. Even if the bytes where converted using the
advertised charest correctly into UTF-8 based strings, a HTML part may
specify its encoding within the document. In this case, clients that
want to work with this content cannot obtain the original bytes. The
`BodyContent` allows to capture the advertised charset as well.