Always use ascii output from jsoup in BodyClean
Jsoup doesn't support outputting the same html as given (which is a
feature, otherwise it couldn't fix broken input). It will always
translate html entities into their corresponding code points if the
encoding supports it ([0]). That means, jsoup translates ascii content
like `&uuml;` into `ü` or `&ndash;` into `–` and so the original
content cannot be retrived afterwards.

The problem is with mails that contain html parts that specify an
UTF-8 encoding in the document `<head>`, are actually ascii only and
use entities like for example `&uuml;`. This is fine, of course, since
utf-8 contains ascii. Then the mail part declares a charset of
iso8859-1 and a 7bit transfer encoding. The html content is just
ascii, so this also works with iso8859-1. It then breaks with the
output of jsoup when using the declared encoding of the mime part
"iso8859-1". Jsoup converts entities like `&uuml;` to its iso8859-1
bytes. This will break the resulting html, because the html header
still specifies an utf8 encoding but now contains bytes for iso8859-1.
Before, emil used the declared encoding of the mime part, which would
work here, too, if the content would be returned as is (ascii). The
workaround is to tell jsoup to always return ascii content. This would
now create really ugly (but valid) html for something like cyrillic
content, though.

[0] https://github.com/jhy/jsoup/pull/56 I think this is related