Avoid calling #getContent unconditionally
The Part#getContent may do a lot of things (depending on registered
data handlers). For example, for an image part, the image is read into
a java.awt.Image and therefore involves the ImageIO plugin system.
This then has a memory and performance implications, obviously, and is
not necessary here. This is now explicitly only called, when the
content type is known.